🩹 fix Duplicate Subscriptions for LanguageComponent & HiveChangedEvent
